Services/FakePSP/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
httpx
python-brcode
qrcode[pil]